LayoutTests/imported/w3c/web-platform-tests/custom-elements/registries/Element-innerHTML-expected.txt


PASS innerHTML on a disconnected element should use the scoped registry it was created with
PASS nested descendants in innerHTML on a disconnected element should use the scoped registry the element was created with
PASS innerHTML on a disconnected element should use the scoped registry it was created with when parsing a simple HTML
PASS innerHTML on an inserted element should continue to use the scoped registry it was created with
FAIL nested descendants in innerHTML should use the null registry when the container element has null registry assert_equals: expected null but got object "[object CustomElementRegistry]"
FAIL insertAdjacentHTML should use the element's registry even when the registry is null assert_equals: expected null but got object "[object CustomElementRegistry]"
